requirements.txt
fastapi
uvicorn
SQLAlchemy
opencv-python
pyzmq
async-exit-stack
async-generator
pandas
sse-starlette
aiofiles

